* Add LoadRemoteCollectionMixin to handle loading remote collections. Since this code got longer replacing jquery, it makes sense to move it to a mixin since it is repeated.
