+++ app/controllers/items_controller.rb
@@ -13,10 +13,6 @@ class ItemsController < ApplicationController
     end
   end
 
-  def show
-    @items_url = Rails.configuration.honeycomb_url + "/v1/items/" + params[:id]
-  end
-
   private
   def set_collections_url
     @collections_url = Rails.configuration.honeycomb_url + "/v1/collections/" + params[:collection_id]
